OGM-770 Moving TX-related classes of compensation API to a more sensible place
