+++ core/src/main/java/org/hibernate/ogm/compensation/impl/ErrorHandlerEnabledTransactionCoordinatorBuilder.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.transaction.errorhandler.impl;
+package org.hibernate.ogm.compensation.impl;
 
 import org.hibernate.ConnectionAcquisitionMode;
 import org.hibernate.ConnectionReleaseMode;
